requirements.txt
Django==2.0.3
django-environ==0.4.4
Pillow==5.0.0
pytz==2018.3
six==1.11.0
